rules.tpl: support $DEB_BUILD_OPTIONS' nodoc

+++ templates/debian/rules.tpl
@@ -9,8 +9,10 @@ export {{key}}={{value}}{% endfor %}
 
 override_dh_auto_build:
 	dh_auto_build
+ifeq (,$(filter nodoc,$(DEB_BUILD_OPTIONS)))
 	cd {{docs.sphinx_dir}} && \
 	PYTHONPATH=$(CURDIR) http_proxy='http://127.0.0.1:9/' https_proxy='https://127.0.0.1:9/' \
 	sphinx-build -N -E -T -b html . $(CURDIR)/.pybuild/docs/html/
 	rm -rf $(CURDIR)/.pybuild/docs/html/.doctrees
+endif
 {% endif %}
